// src/api/userInfoResponse.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * RunRush API
 * 러닝 소셜 플랫폼 RunRush의 REST API 문서
 * OpenAPI spec version: v1.0.0
 */
import type { UserInfoResponseLevel } from './userInfoResponseLevel';

export interface UserInfoResponse {
  userId?: number;
  email?: string;
  name?: string;
  profileImage?: string;
  nickname?: string;
  level?: UserInfoResponseLevel;
  experiencePoints?: number;
}