// src/api/popularRecommendedCourseResponse.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * RunRush API
 * 러닝 소셜 플랫폼 RunRush의 REST API 문서
 * OpenAPI spec version: v1.0.0
 */

export interface PopularRecommendedCourseResponse {
  courseId?: number;
  courseTitle?: string;
  creatorName?: string;
  totalDistance?: number;
  totalCompletionCount?: number;
  uniqueRunnerCount?: number;
  averagePace?: number;
}