// src/api/recommendedCourseListResponse.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * RunRush API
 * 러닝 소셜 플랫폼 RunRush의 REST API 문서
 * OpenAPI spec version: v1.0.0
 */

export interface RecommendedCourseListResponse {
  id?: number;
  userName?: string;
  title?: string;
  imageUrl?: string;
  description?: string;
  endLocationName?: string;
  totalDistance?: number;
  isBookmarked?: boolean;
}